rag-course/03-rag-vanilla/requirements.txt

sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.2
openai==1.45.0
